pyproject.toml
[tool.poetry]
name = "evid"
version = "0.1.0"
description = "document logger, adder of metadata"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "evid", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
arrow = "^1.3.0"
pymupdf = "^1.25.2"
pypdf = "^3.0.1"
pyqt6 = "^6.7.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-qt = "^4.2.0"  # Added for GUI testing

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
evid = "evid.gui.main:main"
